+++ test/test_complex_logic.py
@@ -90,9 +90,9 @@ class TestComplexClient(object):
             assert c.open_outbound_streams == expected_outbound_streams
 
             # Pushed streams can only be closed remotely.
-            f = frame_factory.build_headers_frame(
+            f = frame_factory.build_data_frame(
                 stream_id=stream_id+1,
-                headers=self.example_response_headers,
+                data=b'the content',
                 flags=['END_STREAM'],
             )
             c.receive_data(f.serialize())
@@ -550,7 +550,7 @@ class TestContinuationFramesPushPromise(object):
 
         f = frame_factory.build_goaway_frame(
             last_stream_id=0,
-            error_code=h2.errors.PROTOCOL_ERROR,
+            error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR,
         )
         assert c.data_to_send() == f.serialize()