+++ test/test_invalid_frame_sequences.py
@@ -127,7 +127,7 @@ class TestInvalidFrameSequences(object):
             c.receive_data(invalid_window_update_frame)
 
         expected_frame = frame_factory.build_goaway_frame(
-            last_stream_id=0, error_code=h2.errors.FRAME_SIZE_ERROR
+            last_stream_id=0, error_code=h2.errors.ErrorCodes.FRAME_SIZE_ERROR
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -244,8 +244,8 @@ class TestInvalidFrameSequences(object):
             c.receive_data(f.serialize())
 
         assert e.value.error_code == (
-            h2.errors.FLOW_CONTROL_ERROR if 0x4 in settings else
-            h2.errors.PROTOCOL_ERROR
+            h2.errors.ErrorCodes.FLOW_CONTROL_ERROR if 0x4 in settings else
+            h2.errors.ErrorCodes.PROTOCOL_ERROR
         )
 
     def test_invalid_frame_headers_are_protocol_errors(self, frame_factory):
@@ -303,7 +303,7 @@ class TestInvalidFrameSequences(object):
         event = events[0]
         assert isinstance(event, h2.events.StreamReset)
         assert event.stream_id == 1
-        assert event.error_code == h2.errors.STREAM_CLOSED
+        assert event.error_code == h2.errors.ErrorCodes.STREAM_CLOSED
         assert not event.remote_reset
 
     def test_one_one_stream_reset(self, frame_factory):
@@ -338,7 +338,7 @@ class TestInvalidFrameSequences(object):
         event = events[0]
         assert isinstance(event, h2.events.StreamReset)
         assert event.stream_id == 1
-        assert event.error_code == h2.errors.STREAM_CLOSED
+        assert event.error_code == h2.errors.ErrorCodes.STREAM_CLOSED
         assert not event.remote_reset
 
     @pytest.mark.parametrize('value', ['', 'twelve'])
@@ -360,7 +360,7 @@ class TestInvalidFrameSequences(object):
 
         expected_frame = frame_factory.build_goaway_frame(
             last_stream_id=1,
-            error_code=h2.errors.PROTOCOL_ERROR
+            error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -384,7 +384,33 @@ class TestInvalidFrameSequences(object):
 
         expected_frame = frame_factory.build_goaway_frame(
             last_stream_id=0,
-            error_code=h2.errors.PROTOCOL_ERROR
+            error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR
+        )
+        assert c.data_to_send() == expected_frame.serialize()
+
+    def test_invalid_push_promise_data_protocol_error(self, frame_factory):
+        """
+        If an invalid header block is received on a PUSH_PROMISE, we raise a
+        ProtocolError.
+        """
+        c = h2.connection.H2Connection(client_side=True)
+        c.initiate_connection()
+        c.send_headers(stream_id=1, headers=self.example_request_headers)
+        c.clear_outbound_data_buffer()
+
+        f = frame_factory.build_push_promise_frame(
+            stream_id=1,
+            promised_stream_id=2,
+            headers=self.example_request_headers
+        )
+        f.data = b'\x00\x00\x00\x00'
+
+        with pytest.raises(h2.exceptions.ProtocolError):
+            c.receive_data(f.serialize())
+
+        expected_frame = frame_factory.build_goaway_frame(
+            last_stream_id=0,
+            error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR
         )
         assert c.data_to_send() == expected_frame.serialize()
 
@@ -424,7 +450,7 @@ class TestInvalidFrameSequences(object):
 
         expected_frame = frame_factory.build_goaway_frame(
             last_stream_id=2,
-            error_code=h2.errors.PROTOCOL_ERROR
+            error_code=h2.errors.ErrorCodes.PROTOCOL_ERROR
         )
         assert c.data_to_send() == expected_frame.serialize()